aggregating invoke type utilities
